+++ src/sl_tg01id.cc
@@ -128,7 +128,7 @@ DEFUN_DLD (__sl_tg01id__, args, nargout, "Slicot TG01ID Release 5.0")
             error ("__sl_tg01id__: exception in SLICOT subroutine TG01ID");
             
         if (info != 0)
-            error ("__sl_tg01id__: TG01ID returned info = %d", info);
+            error ("__sl_tg01id__: TG01ID returned info = %d", static_cast<int> (info));
 
         // resize
         a.resize (n, n);
